fix loading bug and move to feed-display directory
