+++ src/components/feed-display/FeedItemDisplay.vue
@@ -101,9 +101,9 @@ import CloseIcon from 'vue-material-design-icons/Close.vue'
 import NcActions from '@nextcloud/vue/dist/Components/NcActions.js'
 import NcActionButton from '@nextcloud/vue/dist/Components/NcActionButton.js'
 
-import { Feed } from '../types/Feed'
-import { FeedItem } from '../types/FeedItem'
-import { ACTIONS, MUTATIONS } from '../store'
+import { Feed } from '../../types/Feed'
+import { FeedItem } from '../../types/FeedItem'
+import { ACTIONS, MUTATIONS } from '../../store'
 
 export default Vue.extend({
 	name: 'FeedItemDisplay',
