+++ src/components/routes/Starred.vue
@@ -20,7 +20,7 @@ import { mapState } from 'vuex'
 
 import NcCounterBubble from '@nextcloud/vue/dist/Components/NcCounterBubble.js'
 
-import FeedItemDisplayList from '../FeedItemDisplayList.vue'
+import FeedItemDisplayList from '../feed-display/FeedItemDisplayList.vue'
 
 import { FeedItem } from '../../types/FeedItem'
 import { ACTIONS, MUTATIONS } from '../../store'
