+++ src/types/MutationTypes.ts
@@ -18,4 +18,5 @@ export const FEED_ITEM_MUTATION_TYPES = {
 
 	SET_FETCHING: 'SET_FETCHING',
 	SET_ALL_LOADED: 'SET_ALL_LOADED',
+	SET_LAST_ITEM_LOADED: 'SET_LAST_ITEM_LOADED',
 }
